refactor: rename style engine base to stylesheet engine
