+++ modules/graph-layers/src/layers/graph-layer.ts
@@ -303,7 +303,10 @@ export class GraphLayer extends CompositeLayer<GraphLayerProps> {
     });
   }
 
-  private _createStyleEngine(style: GraphStylesheet, context: string): GraphStylesheetEngine | null {
+  private _createStylesheetEngine(
+    style: GraphStylesheet,
+    context: string
+  ): GraphStylesheetEngine | null {
     try {
       return new GraphStylesheetEngine(style, {
         stateUpdateTrigger: (this.state.interactionManager as any).getLastInteraction()
@@ -804,7 +807,7 @@ export class GraphLayer extends CompositeLayer<GraphLayerProps> {
           warn(`GraphLayer: Invalid node type "${style.type}".`);
           return null;
         }
-        const stylesheet = this._createStyleEngine(
+        const stylesheet = this._createStylesheetEngine(
           restStyle as unknown as GraphStylesheet,
           `node stylesheet "${style.type}"`
         );
@@ -858,7 +861,7 @@ export class GraphLayer extends CompositeLayer<GraphLayerProps> {
       .filter(Boolean)
       .flatMap((style, idx) => {
         const {decorators, data = (edges) => edges, visible = true, ...restEdgeStyle} = style;
-        const stylesheet = this._createStyleEngine(
+        const stylesheet = this._createStylesheetEngine(
           {
             type: 'edge',
             ...restEdgeStyle
@@ -893,7 +896,7 @@ export class GraphLayer extends CompositeLayer<GraphLayerProps> {
               warn(`GraphLayer: Invalid edge decorator type "${decoratorStyle.type}".`);
               return null;
             }
-            const decoratorStylesheet = this._createStyleEngine(
+            const decoratorStylesheet = this._createStylesheetEngine(
               decoratorStyle as unknown as GraphStylesheet,
               `edge decorator stylesheet "${decoratorStyle.type}"`
             );
@@ -994,7 +997,7 @@ export class GraphLayer extends CompositeLayer<GraphLayerProps> {
       );
     }
 
-    const collapsedMarkerStylesheet = this._createStyleEngine(
+    const collapsedMarkerStylesheet = this._createStylesheetEngine(
       {
         type: 'marker',
         fill: [64, 96, 192, 255],
@@ -1047,7 +1050,7 @@ export class GraphLayer extends CompositeLayer<GraphLayerProps> {
       );
     }
 
-    const expandedMarkerStylesheet = this._createStyleEngine(
+    const expandedMarkerStylesheet = this._createStylesheetEngine(
       {
         type: 'marker',
         fill: [64, 96, 192, 255],
